backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
yfinance==0.2.33
python-multipart==0.0.6
